Fixing waste of tokens on browser open, force new rules on extension update, automatic reload after one-off double spend (#16)
This PR fixes an important and a minor bugs in the extensions, and improves UX:

# Bug

Before this patch, when opening the browser with the extension enabled, one token per endpoint (about 50) would be discarded automatically. This is now resolved.

# Minor bug

The way that the header rules were being applied for the /html/* search pages, was causing the wrong header being sent due to confusion between /html and /html/*.

# UX Feature 1

After extension update, new header rules were not being automatically applied and the user had to manually disable and re-enable PP mode. This is now done automatically.

# UX Feature 2

Sometimes due to a race-condition or a header rule update, the user may encounter an error message about a token having been already spent. With this update, we check if this error had already happened in the prior minute. If not, then the page displaying the error is reloaded using a new token, which likely resolves the problem. If the error had just happened (eg, if after the page reload the problem is not resolved), the new error page is left on display.